[SHLK-296][SHLK-430] Add mtp-responder and changable gadget mode USB using usbfn.sh, default RNDIS_MTP

+++ recipes-connectivity/gadgetload/files/gadget.sh
@@ -7,8 +7,15 @@ udc_device=`ls /sys/class/udc/ | head -n1`
 product="FLIR Camera" # TODO: Populate with product name
 manufacturer="FLIR Systems"
 vendor_id="0x09CB"
-product_id="0x1002"
+product_id=""
+product_id_rndis="0x1002"
+product_id_mtp="0x100A"
+product_id_rndis_mtp="0x100C"
+
 serial="0" # TODO: Should be device serial number?
+
+usbmode="RNDIS_MTP" # RNDIS_MTP | MTP | RNDIS
+usbmode_default="RNDIS_MTP"
 
 device_revision="0x1004"    # Revision, increment this during development when making
                             # changes so that OS enumerates new device instead of using
@@ -73,6 +80,16 @@ usage() {
 }
 
 config_load() {
+    # Check usbmode config and vaildate
+    if [ -f "/etc/usbmode" ] ; then
+        usbmode=`cat /etc/usbmode`
+        if [ ! $usbmode = "MTP" ] && [ ! $usbmode = "RNDIS" ] && [ ! $usbmode = "RNDIS_MTP" ] ; then
+            usbmode="${usbmode_default}"
+        fi
+    else
+        usbmode="${usbmode_default}"
+    fi
+
     if [ ! -d "${gadget_root}" ]; then
         modprobe libcomposite
     fi
@@ -84,6 +101,15 @@ config_load() {
 
     mkdir -p ${gadget_path}
     cd ${gadget_path}
+
+    if [ $usbmode = "RNDIS" ] ; then
+        product_id="${product_id_rndis}"
+    elif [ $usbmode = "MTP" ] ; then
+        product_id="${product_id_mtp}"
+    else
+        product_id="${product_id_rndis_mtp}"
+    fi
+
 
     echo "${usb_version}"       > bcdUSB
     echo "${vendor_id}"         > idVendor
@@ -97,49 +123,65 @@ config_load() {
     echo "${manufacturer}"  > strings/0x409/manufacturer
     echo "${product}"       > strings/0x409/product
     echo "${serial}"        > strings/0x409/serialnumber
-
-    # Config 1 is for RNDIS
+    
+    # Config 1
     mkdir -p configs/c.1
     echo "${usb_attr}"      > configs/c.1/bmAttributes
     echo "${usb_max_power}" > configs/c.1/MaxPower
     mkdir -p configs/c.1/strings/0x409
-    echo "RNDIS" > configs/c.1/strings/0x409/configuration
 
-    # This is together with the ms additions in the rndis function
-    # is to make windows detect the RNDIS device and choose the 6.0
-    # RNDIS driver, even if FLIR Device Drivers are not installed.
-    echo "1"                    > os_desc/use
-    echo "${ms_vendor_code}"    > os_desc/b_vendor_code
-    echo "${ms_qw_sign}"        > os_desc/qw_sign
+    if [ $usbmode = "RNDIS_MTP" ] || [ $usbmode = "RNDIS" ] ; then
+        echo "RNDIS" > configs/c.1/strings/0x409/configuration
+        # This is together with the ms additions in the rndis function
+        # is to make windows detect the RNDIS device and choose the 6.0
+        # RNDIS driver, even if FLIR Device Drivers are not installed.
+        echo "1"                    > os_desc/use
+        echo "${ms_vendor_code}"    > os_desc/b_vendor_code
+        echo "${ms_qw_sign}"        > os_desc/qw_sign
 
-    # Create the RNDIS function
+        # Create the RNDIS function
 
-    # the first digit in the first number of the MAC address is important,
-    # the two first bits of that digit, b10 (d2) means it is reserved to
-    # locally assigned unicast adresses to avoid clashes with existing
-    # "external" MAC adresses
-    mac_base=`get_base_mac_addr`
-    mac_dev="02${mac_base}"
-    mac_host="12${mac_base}"
+        # the first digit in the first number of the MAC address is important,
+        # the two first bits of that digit, b10 (d2) means it is reserved to
+        # locally assigned unicast adresses to avoid clashes with existing
+        # "external" MAC adresses
+        mac_base=`get_base_mac_addr`
+        mac_dev="02${mac_base}"
+        mac_host="12${mac_base}"
 
-    mkdir functions/rndis.usb0
-    echo "${mac_dev}"           > functions/rndis.usb0/dev_addr
-    echo "${mac_host}"          > functions/rndis.usb0/host_addr
-    echo "${ms_compat_id}"      > functions/rndis.usb0/os_desc/interface.rndis/compatible_id
-    echo "${ms_subcompat_id}"   > functions/rndis.usb0/os_desc/interface.rndis/sub_compatible_id
+        mkdir functions/rndis.usb0
+        echo "${mac_dev}"           > functions/rndis.usb0/dev_addr
+        echo "${mac_host}"          > functions/rndis.usb0/host_addr
+        echo "${ms_compat_id}"      > functions/rndis.usb0/os_desc/interface.rndis/compatible_id
+        echo "${ms_subcompat_id}"   > functions/rndis.usb0/os_desc/interface.rndis/sub_compatible_id
 
-    # Link everything up and bind the USB device
-    ln -s functions/rndis.usb0 configs/c.1
+        # Link everything up and bind the USB device
+        ln -s functions/rndis.usb0 configs/c.1
+    fi
+
+    if [ $usbmode = "RNDIS_MTP" ] || [ $usbmode = "MTP" ] ; then
+        # create and link mtp
+        mkdir -p functions/ffs.umtp
+        ln -s functions/ffs.umtp configs/c.1
+        # Mount and start MTP
+        mkdir -p /dev/ffs-umtp
+        mount -t functionfs umtp /dev/ffs-umtp
+        umtprd &
+    fi
+
     ln -s configs/c.1 os_desc
+
+    sleep 1
 
     echo "${udc_device}" > UDC
 
-    # Set ip adress to default adress
-    sleep 1
-    ifconfig usb0 `get_default_usb_ip_addr`
+    if [ $usbmode = "RNDIS_MTP" ] || [ $usbmode = "RNDIS" ] ; then
+        # Set ip adress to default adress
+        ifconfig usb0 `get_default_usb_ip_addr`
 
-    # Signal to fis to update RNDIS status
-    killall -USR1 fis || true
+        # Signal to fis to update RNDIS status
+        killall -USR1 fis || true
+    fi
 
     echo "Loaded: Done."
 }
@@ -162,8 +204,10 @@ config_unload() {
     # remove everything in reverse from creation order in config_load()
     remove_if_exists ${gadget_path}/os_desc/c.1
     remove_if_exists ${gadget_path}/configs/c.1/rndis.usb0
+    remove_if_exists ${gadget_path}/configs/c.1/ffs.umtp
 
     remove_if_exists ${gadget_path}/functions/rndis.usb0
+    remove_if_exists ${gadget_path}/functions/ffs.umtp
     remove_if_exists ${gadget_path}/configs/c.1/strings/0x409
     remove_if_exists ${gadget_path}/configs/c.1
 
@@ -173,6 +217,10 @@ config_unload() {
 
     # Signal to fis to update RNDIS status
     killall -USR1 fis
+    killall umtprd
+    
+    umount -l /dev/ffs-umtp
+    remove_if_exists /dev/ffs-umtp
 
     echo "Unloaded: Done."
 }
